+++ background/lib/utils/index.ts
@@ -1,24 +1,8 @@
 import { utils } from "ethers"
-import HDKeyring, { normalizeHexAddress } from "@tallyho/hd-keyring"
-import {
-  AugmentedLimitOrder,
-  BaseLimitOrder,
-  HexString,
-  KeeperDAOLimitOrder,
-} from "../../types"
+import { normalizeHexAddress } from "@tallyho/hd-keyring"
+import { HexString } from "../../types"
 import { EVMNetwork } from "../../networks"
-import {
-  ETHEREUM,
-  ROPSTEN,
-  RINKEBY,
-  GOERLI,
-  KOVAN,
-  ZEROEX_DOMAIN_DEFAULTS,
-} from "../../constants"
-
-import * as ethjs from "ethereumjs-util"
-import { ethers } from "ethers"
-import axios from "axios"
+import { ETHEREUM, ROPSTEN, RINKEBY, GOERLI, KOVAN } from "../../constants"
 
 export function normalizeEVMAddress(address: string | Buffer): HexString {
   return normalizeHexAddress(address)
@@ -121,97 +105,4 @@ export function getEthereumNetwork(): EVMNetwork {
 
 export function truncateAddress(address: string): string {
   return `${address.slice(0, 6)}...${address.slice(-5)}`
-}
-
-const augmentLimitOrder = async (
-  unformattedLimitOrder: BaseLimitOrder,
-  orderDetails: KeeperDAOOrderDetails
-): Promise<AugmentedLimitOrder> => {
-  return {
-    makerToken: unformattedLimitOrder.makerToken.toLowerCase(),
-    takerToken: unformattedLimitOrder.takerToken.toLowerCase(),
-    makerAmount: unformattedLimitOrder.makerAmount.toString().toLowerCase(),
-    takerAmount: unformattedLimitOrder.takerAmount.toString().toLowerCase(),
-    maker: unformattedLimitOrder.maker.toLowerCase(),
-    expiry: unformattedLimitOrder.expiry.toLowerCase(),
-    taker: orderDetails.taker,
-    txOrigin: orderDetails.txOrigin,
-    pool: orderDetails.pool,
-    salt: String(Date.now()).toLowerCase(),
-  }
-}
-
-interface KeeperDAOOrderDetails {
-  verifyingContract: string
-  chainId: string
-  txOrigin: string
-  taker: string
-  pool: string
-}
-interface KeeperDAOInfoResponse {
-  result: {
-    orderDetails: KeeperDAOOrderDetails
-  }
-}
-
-export const signKeeperDaoLimitOrder = async (
-  order: BaseLimitOrder,
-  keyring: HDKeyring
-): Promise<KeeperDAOLimitOrder> => {
-  const keeperDAOInfo = await axios.get<KeeperDAOInfoResponse>(
-    "https://hidingbook.keeperdao.com/api/v1/info"
-  )
-
-  const { orderDetails } = keeperDAOInfo.data.result
-  // The data to sign
-  const normalizedLimitOrder = await augmentLimitOrder(order, orderDetails)
-
-  // 0x Domain
-  const domain = {
-    ...ZEROEX_DOMAIN_DEFAULTS,
-    verifyingContract: orderDetails.verifyingContract,
-  }
-
-  // The named list of all type definitions
-  const types = {
-    RfqOrder: [
-      { type: "address", name: "makerToken" },
-      { type: "address", name: "takerToken" },
-      { type: "uint128", name: "makerAmount" },
-      { type: "uint128", name: "takerAmount" },
-      { type: "address", name: "maker" },
-      { type: "address", name: "taker" },
-      { type: "address", name: "txOrigin" },
-      { type: "bytes32", name: "pool" },
-      { type: "uint64", name: "expiry" },
-      { type: "uint256", name: "salt" },
-    ],
-  }
-
-  const signed = await keyring.signTypedData(
-    normalizedLimitOrder.maker,
-    domain,
-    types,
-    normalizedLimitOrder as unknown as Record<string, unknown>
-  )
-
-  const vals = ethjs.fromRpcSig(signed)
-
-  ethers.utils.hexlify(vals.r)
-
-  const signature = {
-    signatureType: 2,
-    r: `${ethers.utils.hexlify(vals.r)}`,
-    s: `${ethers.utils.hexlify(vals.s)}`,
-    v: vals.v,
-  }
-
-  return {
-    ...normalizedLimitOrder,
-    chainId: 1,
-    verifyingContract: orderDetails.verifyingContract,
-    signature: {
-      ...signature,
-    },
-  }
 }
