+++ background/main.ts
@@ -4,12 +4,7 @@ import { configureStore, isPlain, Middleware } from "@reduxjs/toolkit"
 import devToolsEnhancer from "remote-redux-devtools"
 import { PermissionRequest } from "@tallyho/provider-bridge-shared"
 
-import {
-  decodeJSON,
-  encodeJSON,
-  getEthereumNetwork,
-  signKeeperDaoLimitOrder,
-} from "./lib/utils"
+import { decodeJSON, encodeJSON, getEthereumNetwork } from "./lib/utils"
 
 import {
   BaseService,
@@ -24,7 +19,7 @@ import {
   ServiceCreatorFunction,
 } from "./services"
 
-import { BaseLimitOrder, KeeperDAOLimitOrder, KeyringTypes } from "./types"
+import { KeyringTypes } from "./types"
 import { EIP1559TransactionRequest, SignedEVMTransaction } from "./networks"
 
 import rootReducer from "./redux-slices"
@@ -66,6 +61,7 @@ import {
 } from "./redux-slices/dapp-permission"
 import logger from "./lib/logger"
 import axios from "axios"
+import { BaseLimitOrder, sendKeeperDaoLimitOrder } from "./lib/keeper-dao"
 
 // This sanitizer runs on store and action data before serializing for remote
 // redux devtools. The goal is to end up with an object that is directly
@@ -400,15 +396,11 @@ export default class Main extends BaseService<never> {
       async (transaction: BaseLimitOrder) => {
         logger.log("Got Limit Order Request!")
         try {
-          const signature = await this.keyringService.signLimitOrder(
-            transaction
+          const keeperDaoResponse = await sendKeeperDaoLimitOrder(
+            transaction,
+            this.keyringService
           )
-
-          const res = await axios.post(
-            "https://hidingbook.keeperdao.com/api/v1/orders",
-            [signature]
-          )
-          logger.log("Limit Order Posted Successfully!", res.data.message)
+          logger.log("Limit Order Posted Successfully!", keeperDaoResponse)
         } catch (e) {
           logger.error(e)
           throw e
