+++ background/services/keyring/index.ts
@@ -2,11 +2,7 @@ import { parse as parseRawTransaction } from "@ethersproject/transactions"
 
 import HDKeyring, { SerializedHDKeyring } from "@tallyho/hd-keyring"
 
-import {
-  normalizeEVMAddress,
-  getEthereumNetwork,
-  signKeeperDaoLimitOrder,
-} from "../../lib/utils"
+import { normalizeEVMAddress, getEthereumNetwork } from "../../lib/utils"
 import { ServiceCreatorFunction, ServiceLifecycleEvents } from "../types"
 import { getEncryptedVaults, writeLatestEncryptedVault } from "./storage"
 import {
@@ -15,17 +11,13 @@ import {
   encryptVault,
   SaltedKey,
 } from "./encryption"
-import {
-  BaseLimitOrder,
-  HexString,
-  KeeperDAOLimitOrder,
-  KeyringTypes,
-  UNIXTime,
-} from "../../types"
+import { HexString, KeyringTypes, UNIXTime } from "../../types"
 import { EIP1559TransactionRequest, SignedEVMTransaction } from "../../networks"
 import BaseService from "../base"
 import { ETH, MINUTE, ZEROEX_DOMAIN_DEFAULTS } from "../../constants"
 import { ethersTransactionRequestFromEIP1559TransactionRequest } from "../chain/utils"
+
+import { TypedDataDomain, TypedDataField } from "@ethersproject/abstract-signer"
 
 export const MAX_KEYRING_IDLE_TIME = 60 * MINUTE
 export const MAX_OUTSIDE_IDLE_TIME = 60 * MINUTE
@@ -313,27 +305,24 @@ export default class KeyringService extends BaseService<Events> {
     return newKeyring.id
   }
 
-  signLimitOrder = async (
-    limitOrder: BaseLimitOrder
-  ): Promise<KeeperDAOLimitOrder> => {
+  signTypedData = async (
+    address: string,
+    domain: TypedDataDomain,
+    types: Record<string, Array<TypedDataField>>,
+    value: Record<string, unknown>
+  ): Promise<string> => {
     this.requireUnlocked()
 
     // find the keyring using a linear search
     const keyring = this.#keyrings.find((kr) => {
-      return kr
-        .getAddressesSync()
-        .includes(normalizeEVMAddress(limitOrder.maker))
+      return kr.getAddressesSync().includes(normalizeEVMAddress(address))
     })
 
     if (!keyring) {
       throw new Error("No Keyring Found")
     }
 
-    const signedKeeperDAOLimitOrder = await signKeeperDaoLimitOrder(
-      limitOrder,
-      keyring
-    )
-    return signedKeeperDAOLimitOrder
+    return keyring.signTypedData(address, domain, types, value)
   }
 
   /**
