+++ background/types.ts
@@ -50,46 +50,3 @@ export enum KeyringTypes {
   metamaskMnemonic = "mnemonic#metamask",
   singleSECP = "single#secp256k1",
 }
-
-export interface BaseLimitOrder {
-  makerToken: string
-  takerToken: string
-  makerAmount: BigInt
-  takerAmount: BigInt
-  maker: string
-  expiry: string
-}
-
-export interface AugmentedLimitOrder {
-  makerToken: string
-  takerToken: string
-  makerAmount: string
-  takerAmount: string
-  maker: string
-  taker: string
-  txOrigin: string
-  pool: string
-  expiry: string
-  salt: string
-}
-
-export interface KeeperDAOLimitOrder {
-  maker: HexString
-  taker: HexString
-  makerAmount: string
-  takerAmount: string
-  makerToken: HexString
-  takerToken: HexString
-  salt: string
-  expiry: string
-  chainId: number
-  txOrigin: HexString
-  pool: string
-  verifyingContract: HexString
-  signature?: {
-    signatureType: number
-    v: number
-    r: string
-    s: string
-  }
-}
